Add sprint planning and definition of done to manual
Summarise the discussion we had as a team about how we plan cards and
how we know when they're "done".

These are intended as useful checklists to remind us of important things
to check rather than rigid definitions. They should help us rather than
getting in the way, and we should update them or manage things differently
if they don't help in practice.