Changes to the connection instructions.
The section on getting the database endpoint isn't needed because you
can use the cluster endpoint described in the existing section to
connect to the database. Having two sets of instructions to get two
different endpoints when you only need one is confusing.

I've also changed it so that any variables are placed in angle brackets
to make it clear that they're variables.